Spec 0007: inline page-extension traits into base page classes
Inlines the 10 `Extends*` wrapper traits under
`packages/admin/src/Support/Pages/Concerns/` into the five base page
classes that consumed them (`BaseListRecords`, `BaseCreateRecord`,
`BaseEditRecord`, `BaseViewRecord`, `BaseManageRelatedRecords`).
Mirrors the resource-side playbook landed in spec 0005.

- Removes the trait layer; method signatures, visibility, and
  `getDefault*` slot names carry over verbatim.
- Collapses `ExtendsTablePagination` to a single `paginateTableQuery`
  override (drops the `getDefaultPaginationQuery` indirection).
- Deletes the ten trait files and the now-empty `Concerns/` directory.

Small follow-ups bundled in:

- Drops `Lunar\Admin\Support\Concerns\CallsHooks` (a one-line re-export
  of the bridge's `CallsHooks` left behind after spec 0006). All 15
  consumers now import the bridge trait directly. `LunarSetList`
  catches downstream code via `RenameClassRector`.
- Moves spec `0006-filament-bridge-package.md` and `0007-pages-refactor.md`
  into `specs/completed/`.
- Adds `LunarSetList::V1_TO_V2_REMOVED_CLASSES` documenting the ten
  deleted page-concern traits for the upgrade docs.
