+++ packages/admin/src/Filament/Resources/OrderResource/Pages/Components/OrderItemsTable.php
@@ -23,9 +23,9 @@ use Illuminate\Support\HtmlString;
 use Livewire\Attributes\Computed;
 use Lunar\Admin\Filament\Resources\ProductResource\Pages\EditProduct;
 use Lunar\Admin\Livewire\Components\TableComponent;
-use Lunar\Admin\Support\Concerns\CallsHooks;
 use Lunar\Core\Models\ProductVariant;
 use Lunar\Core\Models\Transaction;
+use Lunar\Filament\Support\Concerns\CallsHooks;
 use Lunar\Filament\Tables\Components\KeyValue;
 
 /**
