+++ packages/admin/src/Filament/Resources/OrderResource/Pages/ManageOrder.php
@@ -38,13 +38,13 @@ use Lunar\Admin\Filament\Resources\OrderResource\Pages\Components\OrderItemsTabl
 use Lunar\Admin\Support\Actions\Orders\UpdateStatusAction;
 use Lunar\Admin\Support\Actions\PdfDownload;
 use Lunar\Admin\Support\ActivityLog\Concerns\CanDispatchActivityUpdated;
-use Lunar\Admin\Support\Concerns\CallsHooks;
 use Lunar\Admin\Support\Pages\BaseViewRecord;
 use Lunar\Core\Models\Order;
 use Lunar\Core\Models\Tag;
 use Lunar\Core\Models\Transaction;
 use Lunar\Filament\Forms\Components\Tags as TagsComponent;
 use Lunar\Filament\Infolists\Components\Tags;
+use Lunar\Filament\Support\Concerns\CallsHooks;
 
 /**
  * @property Order $record
