--- packages/admin/src/Support/Concerns/CallsHooks.php
@@ -1,10 +1,0 @@
-<?php
-
-namespace Lunar\Admin\Support\Concerns;
-
-use Lunar\Filament\Support\Concerns\CallsHooks as BridgeCallsHooks;
-
-trait CallsHooks
-{
-    use BridgeCallsHooks;
-}
