--- packages/admin/src/Support/Pages/Concerns/ExtendsFooterWidgets.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Lunar\Admin\Support\Pages\Concerns;
-
-trait ExtendsFooterWidgets
-{
-    protected function getDefaultFooterWidgets(): array
-    {
-        return [];
-    }
-
-    protected function getFooterWidgets(): array
-    {
-        return $this->callLunarHook('footerWidgets', $this->getDefaultFooterWidgets());
-    }
-}
