--- packages/admin/src/Support/Pages/Concerns/ExtendsHeaderActions.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Lunar\Admin\Support\Pages\Concerns;
-
-trait ExtendsHeaderActions
-{
-    protected function getDefaultHeaderActions(): array
-    {
-        return [];
-    }
-
-    protected function getHeaderActions(): array
-    {
-        return $this->callLunarHook('headerActions', $this->getDefaultHeaderActions());
-    }
-}
