--- packages/admin/src/Support/Pages/Concerns/ExtendsHeaderWidgets.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Lunar\Admin\Support\Pages\Concerns;
-
-trait ExtendsHeaderWidgets
-{
-    protected function getDefaultHeaderWidgets(): array
-    {
-        return [];
-    }
-
-    protected function getHeaderWidgets(): array
-    {
-        return $this->callLunarHook('headerWidgets', $this->getDefaultHeaderWidgets());
-    }
-}
