--- packages/admin/src/Support/Pages/Concerns/ExtendsTabs.php
@@ -1,16 +1,0 @@
-<?php
-
-namespace Lunar\Admin\Support\Pages\Concerns;
-
-trait ExtendsTabs
-{
-    protected function getDefaultTabs(): array
-    {
-        return [];
-    }
-
-    public function getTabs(): array
-    {
-        return $this->callLunarHook('getTabs', $this->getDefaultTabs());
-    }
-}
