+++ packages/table-rate-shipping/src/Filament/Resources/ShippingExclusionListResource/Schemas/ShippingExclusionListForm.php
@@ -6,7 +6,7 @@ use Filament\Forms;
 use Filament\Schemas\Components\Component;
 use Filament\Schemas\Components\Section;
 use Filament\Schemas\Schema;
-use Lunar\Admin\Support\Concerns\CallsHooks;
+use Lunar\Filament\Support\Concerns\CallsHooks;
 
 class ShippingExclusionListForm
 {
