+++ packages/table-rate-shipping/src/Filament/Resources/ShippingMethodResource/Schemas/ShippingMethodForm.php
@@ -10,8 +10,8 @@ use Filament\Schemas\Components\Section;
 use Filament\Schemas\Components\Utilities\Get;
 use Filament\Schemas\Schema;
 use Illuminate\Database\Eloquent\Model;
-use Lunar\Admin\Support\Concerns\CallsHooks;
 use Lunar\Core\Facades\Converter;
+use Lunar\Filament\Support\Concerns\CallsHooks;
 
 class ShippingMethodForm
 {
