+++ packages/table-rate-shipping/src/Filament/Resources/ShippingZoneResource/Schemas/ShippingZoneForm.php
@@ -9,9 +9,9 @@ use Filament\Schemas\Components\Section;
 use Filament\Schemas\Components\Utilities\Get;
 use Filament\Schemas\Schema;
 use Illuminate\Database\Eloquent\Model;
-use Lunar\Admin\Support\Concerns\CallsHooks;
 use Lunar\Core\Models\Country;
 use Lunar\Core\Models\State;
+use Lunar\Filament\Support\Concerns\CallsHooks;
 use Lunar\Shipping\Models\Contracts\ShippingZone;
 
 class ShippingZoneForm
