+++ specs/completed/0006-filament-bridge-package.md
